+++ protected/humhub/widgets/SiteLogo.php
@@ -6,14 +6,12 @@ class SiteLogo extends \yii\base\Widget
 {
     public const PLACE_TOP_MENU = 'topMenu';
     public const PLACE_LOGIN = 'login';
-    public const PLACE_EMAIL = 'email';
 
     public string $place = 'topMenu';
     public ?int $maxWidth = null;
     public ?int $maxHeight = null;
     public string $id = 'img-logo';
     public string $class = 'rounded';
-    public string $style = '';
 
     public function init()
     {
@@ -22,10 +20,6 @@ class SiteLogo extends \yii\base\Widget
         if ($this->place === static::PLACE_LOGIN) {
             $this->maxWidth = $this->maxWidth ?? 500;
             $this->maxHeight = $this->maxHeight ?? 250;
-        }
-        if ($this->place === static::PLACE_EMAIL) {
-            $this->maxWidth = $this->maxWidth ?? 600;
-            $this->maxHeight = $this->maxHeight ?? 50;
         }
     }
 
@@ -37,8 +31,6 @@ class SiteLogo extends \yii\base\Widget
             'maxHeight' => $this->maxHeight,
             'id' => $this->id,
             'class' => $this->class,
-            'style' => $this->style,
         ]);
     }
-
 }